Moved bounds checker code into utilities and updated module dosctring
